moved useAmazonPayObjects to helper customhooks folder
