+++ support-frontend/assets/helpers/customHooks/useAmazonPayObjects.ts
